+++ CHANGELOG.md
@@ -1,15 +1,17 @@
 # Changelog for the PDI distribution
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+Please note this is the list for the distribution mechanism. The list for each
+sub-project is located in the sub-project `CHANGELOG.md` file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
 ### Added
 
 ### Changed
-* Now require version 1.0.0 of Zpp minimum
+* Fortran support now requires version 1.0.6 of Zpp minimum
 * Bundle version 1.0.6 of Zpp
 
 ### Deprecated