+++ pdi/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### For users
 
 #### Added
+* Lookup for plugins in paths specified by the `PDI_PLUGIN_PATH` env. variable
+* Install plugins as specified by the `INSTALL_PDIPLUGINDIR` cmake variable
 
 #### Changed
 * Fortran support now requires version 1.0.6 of Zpp minimum
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 * Added expression arithmetic operators
+* The `PDI_DEFAULT_PLUGINDIR` cmake variable provides the default plugin installation path
 
 #### Changed